testimony firstore functions complete

+++ lib/services/cloud_firestore.dart
@@ -2,6 +2,7 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:cop_belgium/main.dart';
 import 'package:cop_belgium/models/fasting_model.dart';
 import 'package:cop_belgium/models/testimony_model.dart';
+import 'package:cop_belgium/utilities/connection_checker.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/cupertino.dart';
@@ -9,22 +10,24 @@ import 'package:flutter/cupertino.dart';
 class CloudFireStore {
   final _firestore = FirebaseFirestore.instance;
   final _user = FirebaseAuth.instance.currentUser;
-
-  // upload testimony  to firestore
-
-  // create function with input testimony object.
-  // upload to tesimonies collection
+  final _connectionChecker = ConnectionChecker();
 
   Future<void> createTestimony({required TestimonyInfo testimony}) async {
     try {
-      if (testimony.userId.isNotEmpty &&
-          testimony.title != null &&
-          testimony.description != null &&
-          testimony.date != null) {
-        final docRef = await _firestore
-            .collection('Testimonies')
-            .add(TestimonyInfo.toMap(map: testimony));
-        docRef.update({'id': docRef.id});
+      bool hasConnection = await _connectionChecker.checkConnection();
+
+      if (hasConnection) {
+        if (testimony.userId.isNotEmpty &&
+            testimony.title != null &&
+            testimony.description != null &&
+            testimony.date != null) {
+          final docRef = await _firestore
+              .collection('Testimonies')
+              .add(TestimonyInfo.toMap(map: testimony));
+          docRef.update({'id': docRef.id});
+        }
+      } else {
+        throw ConnectionChecker.connectionException;
       }
     } catch (e) {
       debugPrint(e.toString() + ' davies');
@@ -32,11 +35,21 @@ class CloudFireStore {
     }
   }
 
-  Future<void> updateTestimony({required TestimonyInfo tInfo}) async {
+  Future<void> updateTestimonyInfo({required TestimonyInfo tInfo}) async {
     try {
-      await _firestore.collection('Testimonies').doc(tInfo.id).update(
-            TestimonyInfo.toMap(map: tInfo),
-          );
+      bool hasConnection = await _connectionChecker.checkConnection();
+
+      // voor dat we de tesitmony aanmaken checken we als we interne connectie hebben.
+      // als we internet connectie hebben update de testimony doc
+      // else throw error no internet connection
+
+      if (hasConnection) {
+        await _firestore.collection('Testimonies').doc(tInfo.id).update(
+              TestimonyInfo.toMap(map: tInfo),
+            );
+      } else {
+        throw ConnectionChecker.connectionException;
+      }
     } on FirebaseException catch (e) {
       debugPrint(e.toString());
       rethrow;
@@ -45,7 +58,13 @@ class CloudFireStore {
 
   Future<void> deleteTestimony({required TestimonyInfo? testimony}) async {
     try {
-      await _firestore.collection('Testimonies').doc(testimony!.id).delete();
+      bool hasConnection = await _connectionChecker.checkConnection();
+
+      if (hasConnection) {
+        await _firestore.collection('Testimonies').doc(testimony!.id).delete();
+      } else {
+        throw ConnectionChecker.connectionException;
+      }
     } on FirebaseException catch (e) {
       debugPrint(e.toString());
       rethrow;
@@ -54,40 +73,35 @@ class CloudFireStore {
 
   Future<void> likeDislikeTestimony({required TestimonyInfo tInfo}) async {
     try {
-      //Create a uniek document id  basen on the testimonyInfo id and the userId.
-      // (this can also be done with the .where() where filter and get only the doc with the tInfo.id and userId)
-      String docRef = tInfo.id.toString() + _user!.uid;
+      bool hasConnection = await _connectionChecker.checkConnection();
+      if (hasConnection) {
+        // Create a uniek document id  basen on the testimonyInfo id and the userId.
+        // (this can also be done with the .where() where filter and get only the doc with the tInfo.id and userId)
+        String docRef = tInfo.id.toString() + _user!.uid;
 
-      // Try to get the document in the likes collection.
-      final docSnap = await getTestimonyLikeDoc(tInfo: tInfo, docRef: docRef);
+        // Try to get the document in the likes collection.
+        final docSnap = await getTestimonyLikeDoc(tInfo: tInfo, docRef: docRef);
 
-      // the user has clicked on the like button:
-      // If the doc does not exits in the Testimony likes collection.
-      // Then it means that the user has not liked the testimony.
-      // So a document is created with a uniek id in the likes collection.
-      if (docSnap.exists == false) {
-        // Liked.
-        createTestimonyLikeDoc(tInfo: tInfo, docRef: docRef);
+        // the user has clicked on the like button:
+        // If the doc does not exits in the Testimony likes collection.
+        // Then it means that the user has not liked the testimony.
+        // So a document is created with a uniek id in the likes collection.
+        if (docSnap.exists == false) {
+          // Liked.
+          createTestimonyLikeDoc(tInfo: tInfo, docRef: docRef);
+        } else {
+          // Disliked.
+          // If doc exist then it means the user has disliked.
+          // So we delete the document in the likes collection.
+          deleteTestimonyLikeDoc(tInfo: tInfo, docRef: docRef);
+        }
       } else {
-        // Disliked.
-        // If doc exist then it means the user has disliked.
-        // So we delete the document in the likes collection.
-        deleteTestimonyLikeDoc(tInfo: tInfo, docRef: docRef);
+        throw ConnectionChecker.connectionException;
       }
     } on FirebaseException catch (e) {
       debugPrint(e.toString());
       rethrow;
     }
-  }
-
-  Future<void> deleteTestimonyLikeDoc(
-      {required TestimonyInfo tInfo, required String docRef}) async {
-    return await FirebaseFirestore.instance
-        .collection('Testimonies')
-        .doc(tInfo.id)
-        .collection('likes')
-        .doc(docRef)
-        .delete();
   }
 
   Future<DocumentSnapshot<Map<String, dynamic>>> getTestimonyLikeDoc({
@@ -121,6 +135,21 @@ class CloudFireStore {
             'userId': _user!.uid,
             'date': currenDate
           }));
+    } on FirebaseException catch (e) {
+      debugPrint(e.toString());
+      rethrow;
+    }
+  }
+
+  Future<void> deleteTestimonyLikeDoc(
+      {required TestimonyInfo tInfo, required String docRef}) async {
+    try {
+      await FirebaseFirestore.instance
+          .collection('Testimonies')
+          .doc(tInfo.id)
+          .collection('likes')
+          .doc(docRef)
+          .delete();
     } on FirebaseException catch (e) {
       debugPrint(e.toString());
       rethrow;
